requirements.txt
pymongo
python-dotenv
sentence-transformers
pymongoarrow
Flask==2.3.2
